Resolve conflicts with overlapping fixes from worktree-5/6. Master
already had equivalent fixes for #545 (Stock non-finite handling) and
#555 (SoftwareProduction cache removal). Kept master's versions plus
our #549 (cursor guard) and #551 (UndoManager timeout) fixes.

+++ courant-engine/src/test/java/systems/courant/sd/model/compile/CompiledModelTest.java
@@ -33,7 +33,7 @@ class CompiledModelTest {
     private Model model;
     private Stock population;
     private CompiledModel compiled;
-    private int[] stepHolder;
+    private long[] stepHolder;
     private double[] dtHolder;
     private AtomicInteger resetCount;
 
@@ -47,7 +47,7 @@ class CompiledModelTest {
         model.addStock(population);
         model.addFlow(growth);
 
-        stepHolder = new int[]{0};
+        stepHolder = new long[]{0};
         dtHolder = new double[]{1.0};
         resetCount = new AtomicInteger(0);
